+++ Assets/GameScene/Scripts/CharacterDetector.cs
@@ -1,14 +1,26 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
 public class CharacterDetector : MonoBehaviour
 {
+    public bool isShowDebugMessages = true;
+
     private void OnTriggerEnter(Collider other)
     {
         if (other.CompareTag(ConstantsStrings.PlayerTag))
         {
+            if (isShowDebugMessages)
+            {
+                Debug.Log("The player entered to the trigger");
+            }
+
             GameManager.Instance.ExtendRoad(true);
+        }
+        else
+        {
+            if (isShowDebugMessages)
+            {
+                Debug.Log("The trigger was activated");
+            }
         }
     }
 }