+++ Assets/TEST/Test_Scripts/Test_Mover.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+public class Test_Mover : MonoBehaviour
+{
+    private void Update()
+    {
+        if (GameManager.Instance.IsPlaing)
+        {
+            transform.Translate(GameManager.Instance.RoadMoovingSpeed * Time.deltaTime);
+        }
+    }
+}