+++ Assets/TEST/Test_Scripts/Test_Road.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+public class Test_Road : MonoBehaviour
+{
+    public Transform beginPoint;
+    public Transform endPoint;
+}